Provide file type flag definitions
These definitions are a bit ugly at this stage, bamType and readType
should be defined against TYPE_BAM and TYPE_READ, but these are defined
in sam.c and so not visible.

Future approach would be to defined as 1<<iota and test that these match
the libbam definitions (purely aesthetic decision).