Use closure to pass data to BamFetchFn earlier
Since we can use a closure (and already did to create the function
passed to bam_fetch) there is no need to take an interface{} and operate
on that.

bam_fetch makes no change to *data and only passes data to func, so the
best way to handle the situation is create the closure outside bamFetch
where the type is known and pass that closure into bamFetch,
disregarding data completely.

e.g. (loosely)
  var c int
  fn := func(_ *bamRecord) int { c++; return 0 }
  ret, err := sf.bamFetch(bi, tid, beg, end, BamFetchFn(fn))