Init hash before bam_get_tid call
The hash is not initialised when created, so we do it when a query is
made (the bam_aux.c function checks that the *hash is NULL before doing
the init, so the cost is minimal).

Also make the samFile free the hash when samClose is called if it is not
free.

Function declarations are required in the CGO header since the init and
destroy hash functions are not publicised.