Open SAM/BAM files by fd
Opening by fd will allow us to use a pipe from an io.Reader/io.Writer so
that more generic communication can be performed. It also allows us to
disentangle ourselves from the use of magic filenames to get stdout and
stdin streams (this is retained for compatibilty within samtools).

Documentation for samopen also updated to include all options and
correct an incorrect statement in the docs. It is not clear why the 'u'
mode is provided since the default compression mode is 0 anyway.